+++ src/clj/rems/db/test_data_helpers.clj
@@ -209,6 +209,7 @@
                                                (:header :label) ""
                                                :date "2002-03-04"
                                                :email "user@example.com"
+                                               :phone-number "+358451110000"
                                                :table (repeat 2 (for [column (:field/columns field)]
                                                                   {:column (:key column) :value field-value}))
                                                :attachment (str attachment)