+++ test/clj/rems/api/test_forms.clj
@@ -189,6 +189,9 @@
                                  {:field/type :email
                                   :field/title localized
                                   :field/optional true}
+                                 {:field/type :phone-number
+                                  :field/title localized
+                                  :field/optional true}
                                  {:field/type :date
                                   :field/title localized
                                   :field/optional true}