+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Changes since v2.16
 - The example-theme now uses the Lato font.
 - The field id can now be changed in the form editor. (#1804)
 - "More info" support for EGA style resources (when :enable-ega flag is set) (#2466)
+- Add phone number to field types in the form. (#2552)
 
 ## v2.16 "Länsiväylä" 2021-02-04